test: custom repository 테스트 분리
